[APP] ANR bug fixed
