+++ app/src/main/java/wannabit/io/cosmostaion/ui/main/chain/cosmos/CosmosActivity.kt
@@ -12,7 +12,6 @@ import wannabit.io.cosmostaion.common.BaseActivity
 import wannabit.io.cosmostaion.common.BaseData
 import wannabit.io.cosmostaion.common.toMoveBack
 import wannabit.io.cosmostaion.data.repository.chain.ProposalRepositoryImpl
-import wannabit.io.cosmostaion.data.repository.tx.TxRepositoryImpl
 import wannabit.io.cosmostaion.data.repository.wallet.WalletRepositoryImpl
 import wannabit.io.cosmostaion.database.Prefs
 import wannabit.io.cosmostaion.databinding.ActivityCosmosBinding
@@ -20,14 +19,11 @@ import wannabit.io.cosmostaion.ui.viewmodel.chain.ProposalViewModel
 import wannabit.io.cosmostaion.ui.viewmodel.chain.ProposalViewModelProviderFactory
 import wannabit.io.cosmostaion.ui.viewmodel.intro.WalletViewModel
 import wannabit.io.cosmostaion.ui.viewmodel.intro.WalletViewModelProviderFactory
-import wannabit.io.cosmostaion.ui.viewmodel.tx.TxViewModel
-import wannabit.io.cosmostaion.ui.viewmodel.tx.TxViewModelProviderFactory
 
 class CosmosActivity : BaseActivity() {
 
     private lateinit var binding: ActivityCosmosBinding
 
-    private lateinit var txViewModel: TxViewModel
     private lateinit var walletViewModel: WalletViewModel
     private lateinit var proposalViewModel: ProposalViewModel
 
@@ -70,12 +66,6 @@ class CosmosActivity : BaseActivity() {
     }
 
     private fun initViewModel() {
-        val txRepository = TxRepositoryImpl()
-        val txViewModelProviderFactory = TxViewModelProviderFactory(txRepository)
-        txViewModel = ViewModelProvider(
-            this, txViewModelProviderFactory
-        )[TxViewModel::class.java]
-
         val walletRepository = WalletRepositoryImpl()
         val walletViewModelProviderFactory = WalletViewModelProviderFactory(walletRepository)
         walletViewModel =
